find_stage_intervals
This folder included the main function: stage_finder. It generates a
workspace file including the target stage intervals, the NIU stage
intervals and the overlapping intervals. The overlapping is determined
by the majority rule. These functions will be used to generate the .mat
files for annotation containing all the waveforms for the specified
stage throughout the sleep study for each subject